+++ domain_project/src/domain/promocode/PromocodeFactory.java
@@ -6,7 +6,7 @@ import domain.product.Product;
  */
 
 public class PromocodeFactory {
-	public static Promocode makePromocode(PromocodeType type, int id, double amount, Product product, int percentage, double saleAmountNeeded) throws Exception{
+	public static Promocode makePromocode(PromocodeType type, int id, double amount, Product product, double percentage, double saleAmountNeeded) throws Exception{
 		Promocode promocode = null;
 		switch(type){
 		case PRODUCTAMOUNT:
